+++ Client.Core.Test/AbstractMockServerTest.cs
@@ -7,7 +7,7 @@ namespace InfluxDB.Client.Core.Test
 {
     public class AbstractMockServerTest : AbstractTest
     {
-        protected FluentMockServer MockServer;
+        protected WireMockServer MockServer;
         protected string MockServerUrl;
 
         [SetUp]
@@ -18,7 +18,7 @@ namespace InfluxDB.Client.Core.Test
                 return;
             }
             
-            MockServer = FluentMockServer.Start(new FluentMockServerSettings
+            MockServer = WireMockServer.Start(new WireMockServerSettings
             {
                 UseSSL = false
             });