+++ Client.Legacy/FluxClient.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
+using System.Text;
 using System.Threading.Tasks;
 using InfluxDB.Client.Core;
 using InfluxDB.Client.Core.Exceptions;
@@ -16,7 +17,7 @@ namespace InfluxDB.Client.Flux
     {
         private readonly LoggingHandler _loggingHandler;
 
-        public FluxClient(FluxConnectionOptions options) : base(new RestClient())
+        public FluxClient(FluxConnectionOptions options) : base(new RestClient(), new FluxResultMapper())
         {
             _loggingHandler = new LoggingHandler(LogLevel.None);
 
@@ -29,7 +30,7 @@ namespace InfluxDB.Client.Flux
             {
                 if (FluxConnectionOptions.AuthenticationType.BasicAuthentication.Equals(options.Authentication))
                 {
-                    var auth = System.Text.Encoding.UTF8.GetBytes(options.Username + ":" + new string(options.Password));
+                    var auth = Encoding.UTF8.GetBytes(options.Username + ":" + new string(options.Password));
                     RestClient.AddDefaultHeader("Authorization", "Basic " + Convert.ToBase64String(auth));
                 }
                 else
@@ -76,7 +77,7 @@ namespace InfluxDB.Client.Flux
         {
             var measurements = new List<T>();
 
-            var consumer = new FluxResponseConsumerPoco<T>((cancellable, poco) => { measurements.Add(poco); });
+            var consumer = new FluxResponseConsumerPoco<T>((cancellable, poco) => { measurements.Add(poco); }, Mapper);
 
             await QueryAsync(query, GetDefaultDialect(), consumer, ErrorConsumer, EmptyAction).ConfigureAwait(false);
 
@@ -185,7 +186,7 @@ namespace InfluxDB.Client.Flux
             Arguments.CheckNotNull(onError, "onError");
             Arguments.CheckNotNull(onComplete, "onComplete");
 
-            var consumer = new FluxResponseConsumerPoco<T>(onNext);
+            var consumer = new FluxResponseConsumerPoco<T>(onNext, Mapper);
 
             return QueryAsync(query, GetDefaultDialect(), consumer, onError, onComplete);
         }