+++ Client.Test/RetryAttemptTest.cs
@@ -142,7 +142,7 @@ namespace InfluxDB.Client.Test
 
         private HttpException CreateException(int retryAfter = 10)
         {
-            var headers = new List<HttpHeader> {new HttpHeader {Name = "Retry-After", Value = retryAfter.ToString()}};
+            var headers = new List<HttpHeader> {new HttpHeader("Retry-After", retryAfter.ToString())};
             var exception = HttpException.Create("", headers, "", HttpStatusCode.TooManyRequests);
             return exception;
         }